feat(results): split participant-voice data into live + crp variants
- Renames `src/data/participant-voice.json` to `participant-voice-crp.json`
  (frozen N=200 CRP dataset).
- Adds `src/data/participant-voice-live.json` (currently a copy of the
  CRP file as a placeholder).
- Each page now reads its own JSON: the CRP page stays frozen, the live
  page will be refreshed from the full Qualtrics dataset by the
  q74-full-analysis workflow (PR #1905).
- Updates the variant intros to make the dataset distinction explicit:
  CRP says "frozen N=200 ... will not change", live says "full TABS V2
  dataset ... refreshes as new respondents are folded in".

The live and crp JSON are byte-identical until #1905 merges and the
workflow runs. A follow-up will commit the real full-dataset numbers
into participant-voice-live.json.
